+++ module/owlib/src/c/ow_1991.c
@@ -79,16 +79,16 @@ static char global_passwd[3][8] = { "", "", "" };
 /* ------- Functions ------------ */
 
 
-static int OW_r_ident( BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn );
-static int OW_w_ident( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn );
-static int OW_w_reset_password( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn );
-static int OW_w_change_password( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn );
-static int OW_r_page( BYTE * data, const size_t size, const size_t offset, const struct parsedname * pn ) ;
-static int OW_w_page( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) ;
-static int OW_r_memory( BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn );
-static int OW_w_memory( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn );
-static int OW_r_subkey( BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn, const int extension );
-static int OW_w_subkey( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn, const int extension );
+static int OW_r_ident( BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn );
+static int OW_w_ident( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn );
+static int OW_w_reset_password( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn );
+static int OW_w_change_password( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn );
+static int OW_r_page( BYTE * data, const size_t size, const off_t offset, const struct parsedname * pn ) ;
+static int OW_w_page( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) ;
+static int OW_r_memory( BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn );
+static int OW_w_memory( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn );
+static int OW_r_subkey( BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn, const int extension );
+static int OW_w_subkey( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn, const int extension );
 
 /* array with magic bytes representing the Copy Scratch operations */
 enum block { ALL=0, IDENT, PASSWORD, DATA };
@@ -157,7 +157,7 @@ static int FS_w_memory( const BYTE *buf, const size_t size, const off_t offset ,
   return 0;
 }
 
-static int OW_w_reset_password( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_w_reset_password( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
     BYTE set_password[3] = { 0x5A, 0x00, 0x00} ;
     char passwd[8];
     char ident[8];
@@ -187,7 +187,7 @@ static int OW_w_reset_password( const BYTE * data , const size_t size , const si
     return 0 ;
 }
 
-static int OW_w_subkey( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn, const int extension ) {
+static int OW_w_subkey( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn, const int extension ) {
     BYTE p[4] = { 0x99, 0x00, 0x00, 0x00} ; // write subkey
     char ident[8];
     struct transaction_log tscratch[] = {
@@ -211,7 +211,7 @@ static int OW_w_subkey( const BYTE * data , const size_t size , const size_t off
     return 0 ;
 }
 
-static int OW_r_subkey( BYTE * data , const size_t size , const size_t offset, const struct parsedname *pn, const int extension ) {
+static int OW_r_subkey( BYTE * data , const size_t size , const off_t offset, const struct parsedname *pn, const int extension ) {
     BYTE p[3] = { 0x66, 0x00, 0x00} ; // read subkey
     struct transaction_log tscratch[] = {
         TRXN_START,
@@ -242,7 +242,7 @@ static int OW_r_subkey( BYTE * data , const size_t size , const size_t offset, c
     return 0 ;
 }
 
-static int OW_r_memory( BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_r_memory( BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
   char all_data[0x40];
   int i, ret, nr_bytes;
   size_t left = size;
@@ -264,7 +264,7 @@ static int OW_r_memory( BYTE * data , const size_t size , const size_t offset, c
   return 0 ;
 }
 
-static int OW_w_memory( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_w_memory( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
   char all_data[0x40];
   int i, ret, nr_bytes;
   size_t left = size;
@@ -293,7 +293,7 @@ static int OW_w_memory( const BYTE * data , const size_t size , const size_t off
 }
 
 // size and offset already bounds checked
-static int OW_r_ident( BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_r_ident( BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
     char all_data[0x40];
     //printf("OW_r_ident\n");
 
@@ -303,7 +303,7 @@ static int OW_r_ident( BYTE * data , const size_t size , const size_t offset, co
     return 0 ;
 }
 
-static int OW_w_ident( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_w_ident( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
     BYTE write_scratch[3] = { 0x96, 0x00, 0x00 } ;
     BYTE copy_scratch[3] = { 0x3C, 0x00, 0x00 } ;
     char all_data[0x40];
@@ -338,7 +338,7 @@ static int OW_w_ident( const BYTE * data , const size_t size , const size_t offs
     return 0 ;
 }
 
-static int OW_w_change_password( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_w_change_password( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
     BYTE write_scratch[3] = { 0x96, 0x00, 0x00 } ;
     BYTE copy_scratch[3] = { 0x3C, 0x00, 0x00 } ;
     char all_data[0x40];
@@ -373,7 +373,7 @@ static int OW_w_change_password( const BYTE * data , const size_t size , const s
     return 0 ;
 }
 
-static int OW_r_page( BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_r_page( BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
     char all_data[0x40];
     int ret ;
 
@@ -387,7 +387,7 @@ static int OW_r_page( BYTE * data , const size_t size , const size_t offset, con
     return 0 ;
 }
 
-static int OW_w_page( const BYTE * data , const size_t size , const size_t offset, const struct parsedname * pn ) {
+static int OW_w_page( const BYTE * data , const size_t size , const off_t offset, const struct parsedname * pn ) {
     BYTE write_scratch[3] = { 0x96, 0xC0, ~(0xC0) } ;
     BYTE copy_scratch[3] = { 0x3C, 0x00, 0x00 } ;
     BYTE all_data[0x40];
